requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.1
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
django-allauth==0.41.0
django-rest-auth
whitenoise